@@ -2,14 +2,13 @@ package tv.slicedlime.tradertest;
 
 import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.context.CommandContext;
-import net.minecraft.command.CommandException;
-import net.minecraft.server.command.CommandManager;
-import net.minecraft.server.command.ServerCommandSource;
-import net.minecraft.server.world.ServerWorld;
-import net.minecraft.text.LiteralText;
-import net.minecraft.world.WanderingTraderManager;
-import net.minecraft.world.gen.Spawner;
-import org.jetbrains.annotations.Nullable;
+import net.minecraft.commands.CommandRuntimeException;
+import net.minecraft.commands.CommandSourceStack;
+import net.minecraft.commands.Commands;
+import net.minecraft.network.chat.Component;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.world.entity.npc.WanderingTraderSpawner;
+import net.minecraft.world.level.CustomSpawner;
 import tv.slicedlime.tradertest.mixin.ServerWorldMixin;
 import tv.slicedlime.tradertest.mixin.WanderingTraderManagerMixin;
 
@@ -17,35 +16,34 @@ import java.util.List;
 
 public class SpawnTraderCommand {
     public static void register(CommandDispatcher dispatcher) {
-        dispatcher.register(CommandManager.literal("spawntrader")
+        dispatcher.register(Commands.literal("spawntrader")
                 .executes(
                     SpawnTraderCommand::run
                 ));
     }
 
-    public static int run(CommandContext<ServerCommandSource> context) {
-        ServerWorld world = context.getSource().getWorld();
-        WanderingTraderManager manager = getWanderingTraderManager(world);
+    public static int run(CommandContext<CommandSourceStack> context) {
+        ServerLevel world = context.getSource().getLevel();
+        WanderingTraderSpawner manager = getWanderingTraderManager(world);
         if (manager != null) {
             for (int i = 0; i < 1000; i++) {
-                boolean success = ((WanderingTraderManagerMixin) manager).trySpawn(world);
+                boolean success = ((WanderingTraderManagerMixin) manager).spawn(world);
                 if (success) {
-                    context.getSource().sendFeedback(new LiteralText("Success"), false);
+                    context.getSource().sendSuccess(Component.literal("Success"), false);
                     return 1;
                 }
             }
-            throw new CommandException(new LiteralText("All spawn attempts failed"));
+            throw new CommandRuntimeException(Component.literal("All spawn attempts failed"));
         } else {
-            throw new CommandException(new LiteralText("No spawner available"));
+            throw new CommandRuntimeException(Component.literal("No spawner available"));
         }
     }
 
-    @Nullable
-    static WanderingTraderManager getWanderingTraderManager(ServerWorld world) {
-        List<Spawner> spawners = ((ServerWorldMixin) world).getSpawners();
-        for (Spawner spawner : spawners) {
-            if (spawner instanceof WanderingTraderManager) {
-                return (WanderingTraderManager) spawner;
+    static WanderingTraderSpawner getWanderingTraderManager(ServerLevel world) {
+        List<CustomSpawner> spawners = ((ServerWorldMixin) world).getCustomSpawners();
+        for (CustomSpawner spawner : spawners) {
+            if (spawner instanceof WanderingTraderSpawner traderSpawner) {
+                return traderSpawner;
             }
         }
         return null;
