@@ -1,16 +1,14 @@
 package tv.slicedlime.tradertest.mixin;
 
-import net.minecraft.server.world.ServerWorld;
-import net.minecraft.world.WanderingTraderManager;
-import net.minecraft.world.gen.Spawner;
-import org.jetbrains.annotations.Nullable;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.world.level.CustomSpawner;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Accessor;
 
 import java.util.List;
 
-@Mixin(ServerWorld.class)
+@Mixin(ServerLevel.class)
 public interface ServerWorldMixin {
     @Accessor
-    List<Spawner> getSpawners();
+    List<CustomSpawner> getCustomSpawners();
 }
