@@ -1,12 +1,12 @@
 package tv.slicedlime.tradertest.mixin;
-
-import net.minecraft.server.world.ServerWorld;
-import net.minecraft.world.WanderingTraderManager;
+;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.world.entity.npc.WanderingTraderSpawner;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Invoker;
 
-@Mixin(WanderingTraderManager.class)
+@Mixin(WanderingTraderSpawner.class)
 public interface WanderingTraderManagerMixin {
-	@Invoker("trySpawn")
-	public boolean trySpawn(ServerWorld world);
+	@Invoker("spawn")
+	public boolean spawn(ServerLevel world);
 }
